ci: adopt ruff 0.16's expanded rule set and clear the findings
Ruff 0.16.0 raised the default rule set from 59 rules to 413. This
adopts the wider set instead of pinning the old one, and clears the 337
findings it raised, in one commit.

What changed in the code is mechanical: import ordering, f-strings over
percent and .format, collection literals over dict() and list() calls,
ClassVar on shared test constants, itertools.pairwise, .items() where a
loop indexed back into its own dict, and context-managed reads in tests
that leaked file handles. No behaviour moves. The suite is 1094 passed,
260 skipped, before and after.

One finding was a real bug. tests/test_read_gate_wire.py evaluated
 with no assert, so the test asserted
nothing. It asserts now.

Five rules are switched off deliberately, each with its reason. Two
date rules are repo-wide in ruff.toml: market dates are naive on
purpose, and attaching a timezone would shift them. The other three are
per-site , the convention this repo already
uses. Broad excepts in the search loops keep a campaign running past
one bad cell, and ValueError is the wire contract the read-gate tests
assert on.

target-version is now declared as py312, matching CI and the local
venv. Undeclared, ruff guessed 3.10 and rewrote a zip() into
itertools.pairwise; the local venv was still 3.9 and failed at runtime.
The venv now matches CI, so the guess and the truth agree.

Line numbers moved, so the citing prose moved with them: 214 textual
file.py:N references across the gap docs and the test plan, plus the
README and tutorial #L anchors, all remapped from the diff and verified
to point at byte-identical code. The notebook is regenerated.

+++ realchains/vol_premium.py
@@ -56,8 +56,9 @@ realistic transaction costs. Phase B (the put side) is the remaining open work.
 from __future__ import annotations
 
 import math
+from collections.abc import Callable
 from datetime import datetime
-from typing import Any, Callable
+from typing import Any
 
 import numpy as np
 import pandas as pd
@@ -1107,13 +1108,13 @@ def run_real_structure_overlay(
             if hedge_mode == 'per_leg_sign':
                 leg = legs[0]
                 lo, hi = (-1.0, 0.0) if leg['right'] == 'put' else (0.0, 1.0)
-                target_hedge = int(round(min(max(leg['delta'], lo), hi) * shares))
+                target_hedge = round(min(max(leg['delta'], lo), hi) * shares)
             else:  # combined: neutralize the net POSITION delta (−Σ sign·delta). For an all-short
                 # structure this equals Σ delta (the old form) bit-for-bit, so the straddle/strangle
                 # pins are unchanged; for a MIXED-sign reversal (short put + long call) it is the only
                 # correct form — Σ delta would read ~0 and leave the +0.5 position delta unhedged.
                 combined = -sum(leg['sign'] * leg['delta'] for leg in legs)
-                target_hedge = int(round(min(max(combined, -1.0), 1.0) * shares))
+                target_hedge = round(min(max(combined, -1.0), 1.0) * shares)
         else:
             target_hedge = 0
         hedge_trade = target_hedge - hedge_shares
